Support Metadata v14 (#206) (#29)
We add support for Metadata v14.
This version introduces some changes. Namely, whereas before
there were three variants of Map (Map, DoubleMap, and nMap),
there is only one Map now, which abstract the previous three.

To follow that design, we also updated our API around checking
the type of a storage entry, now only offering IsPlain() and
IsMap(). For plain types, a Hasher() getter is made available,
and for Maps a Hashers() function alone is provided, which
returns an array of hashers, one hasher per key of the map.

Co-authored-by: Miguel Hervas <miguel.hervas.lazaro@gmail.com>
Co-authored-by: Gkirito <Gkirito@users.noreply.github.com>

Co-authored-by: Nuno Alexandre <hi@nunoalexandre.com>
Co-authored-by: Miguel Hervas <miguel.hervas.lazaro@gmail.com>
Co-authored-by: Gkirito <Gkirito@users.noreply.github.com>